CWE-115481: Update Jenkins Common Plugin https://agile.compuware.com/browse/CWE-115481
