+++ src/main/java/com/compuware/jenkins/common/configuration/CpwrGlobalConfiguration.java
@@ -22,15 +22,11 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.ResourceBundle;
 import java.util.Set;
-import org.apache.commons.lang.StringUtils;
-import org.kohsuke.stapler.QueryParameter;
 import org.kohsuke.stapler.StaplerRequest;
-import com.compuware.jenkins.common.utils.Constants;
 import com.compuware.jenkins.common.utils.NumericStringComparator;
 import hudson.CopyOnWrite;
 import hudson.Extension;
 import hudson.Launcher;
-import hudson.util.FormValidation;
 import hudson.util.ListBoxModel;
 import jenkins.model.GlobalConfiguration;
 import net.sf.json.JSONObject;
@@ -51,6 +47,14 @@ public class CpwrGlobalConfiguration extends GlobalConfiguration
 
 	private String m_topazCLILocationWindows;
 	private String m_topazCLILocationLinux;
+
+    /**
+     * @return the Jenkins managed singleton for the configuration object 
+     */ 
+    public static CpwrGlobalConfiguration get() 
+    { 
+        return GlobalConfiguration.all().get(CpwrGlobalConfiguration.class); 
+    }
 
 	/**
 	 * Constructor.
@@ -127,106 +131,10 @@ public class CpwrGlobalConfiguration extends GlobalConfiguration
 	}
 
 	/**
-	 * Validation for the 'Host:port' text field.
-	 * 
-	 * @param value
-	 *            value passed from the config.jelly "hostPort" field
-	 * 
-	 * @return validation message
-	 */
-	public FormValidation doCheckHostPort(@QueryParameter String value)
-	{
-		FormValidation result;
-
-		String tempValue = StringUtils.trimToEmpty(value);
-		if (tempValue.isEmpty())
-		{
-			result = FormValidation.error(Messages.checkHostPortEmptyError());
-		}
-		else
-		{
-			String[] hostPortParts = StringUtils.split(tempValue, Constants.COLON);
-			if (hostPortParts.length == 2)
-			{
-				String host = StringUtils.trimToEmpty(hostPortParts[0]);
-				String port = StringUtils.trimToEmpty(hostPortParts[1]);
-				result = validateHostPort(host, port);
-			}
-			else if (hostPortParts.length > 2)
-			{
-				result = FormValidation.error(Messages.checkHostPortFormatError());
-			}
-			else
-			{
-				int index = tempValue.indexOf(Constants.COLON);
-				if (index == -1)
-				{
-					result = FormValidation.error(Messages.checkHostPortFormatError());
-				}
-				else if (index == 0)
-				{
-					result = FormValidation.error(Messages.checkHostPortMissingHostError());
-				}
-				else
-				{
-					result = FormValidation.error(Messages.checkHostPortMissingPortError());
-				}
-			}
-		}
-
-		return result;
-	}
-
-	/**
-	 * Validation for the host and port.
-	 * 
-	 * @param host
-	 *            the host value
-	 * @param port
-	 *            the port value
-	 * @return validation message
-	 */
-	private FormValidation validateHostPort(String host, String port)
-	{
-		FormValidation result = FormValidation.ok();
-
-		if (host.isEmpty())
-		{
-			result = FormValidation.error(Messages.checkHostPortMissingHostError());
-		}
-		else if (port.isEmpty())
-		{
-			result = FormValidation.error(Messages.checkHostPortMissingPortError());
-		}
-		else if (!StringUtils.isNumeric(port))
-		{
-			result = FormValidation.error(Messages.checkHostPortInvalidPortError());
-		}
-		return result;
-	}
-
-	/**
-	 * Validation for the host connection name text field.
-	 * 
-	 * @param value
-	 *            value passed from the config.jelly "connectionName" field
-	 * 
-	 * @return validation message
-	 */
-	public FormValidation doCheckConnectionName(@QueryParameter String value)
-	{
-		String tempValue = StringUtils.trimToEmpty(value);
-		if (tempValue.isEmpty())
-		{
-			return FormValidation.error(Messages.checkConnectionNameEmptyError());
-		}
-
-		return FormValidation.ok();
-	}
-
-	/**
 	 * Returns the Topaz Workbench CLI location based on node
 	 * 
+	 * @param launcher
+	 *            launcher for starting a process
 	 * @return CLI location
 	 */
 	public String getTopazCLILocation(Launcher launcher)
