+++ src/main/resources/com/compuware/jenkins/common/configuration/CpwrGlobalConfiguration/config.jelly
@@ -1,40 +1,19 @@
 <?jelly escape-by-default='true'?>
 <j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout"
-         xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:c="/lib/credentials">
-         
-  <f:section title="${%title}">
-  
-      <f:entry title="${%topazWbCliDirectoryWindows}" field="topazCLILocationWindows">
-        <f:textbox name="topazCLILocationWindows" default="C:\Program Files\Compuware\Topaz Workbench CLI"/>
-      </f:entry>
-      <f:entry title="${%topazWbCliDirectoryLinux}" field="topazCLILocationLinux">
-        <f:textbox name="topazCLILocationLinux" default="/opt/Compuware/TopazCLI"/>
-	  </f:entry>
-      
-      <f:entry title="${%hostConnections}" description="${%hostConnectionsDescr}">
-      	<div class="cpwr-section">
-	      <f:repeatable var="hostConn" items="${instance.hostConnections}" add="${%addHostConnection}">
-	        <table width="100%" class="cpwr-hostConnection">
-	          <f:entry title="${%name}" field="connectionName" >
-	            <f:textbox value="${hostConn.getConnectionName()}" checkUrl="'${rootURL}/descriptor/CpwrGlobalConfiguration/checkConnectionName?value='+this.value"/>
-	          </f:entry>
-	          
-	          <f:entry title="${%hostPort}" field="hostPort">
-	            <f:textbox value="${hostConn.getHostPort()}" checkUrl="'${rootURL}/descriptor/CpwrGlobalConfiguration/checkHostPort?value='+this.value"/>
-	          </f:entry>
-	         
-	      	  <f:entry title="${%codePage}" field="codePage">
-	            <f:select default="1047" value="${hostConn.getCodePage()}"/>
-	          </f:entry>
-	      	  
-	          <f:entry title="">
-	            <div align="right">
-	              <f:repeatableDeleteButton value="${%deleteHostConnection}"/>
-	            </div>
-	          </f:entry>
-	        </table>
-	      </f:repeatable>
-      </div>
-    </f:entry>
-  </f:section>
+         xmlns:t="/lib/hudson" xmlns:f="/lib/form">         
+	<f:section title="${%title}">
+    	<f:entry title="${%topazWbCliDirectoryWindows}" field="topazCLILocationWindows">
+        	<f:textbox name="topazCLILocationWindows" default="C:\Program Files\Compuware\Topaz Workbench CLI"/>
+      	</f:entry>
+      	<f:entry title="${%topazWbCliDirectoryLinux}" field="topazCLILocationLinux">
+        	<f:textbox name="topazCLILocationLinux" default="/opt/Compuware/TopazCLI"/>
+	  	</f:entry>
+      	<f:entry title="${%hostConnections}" description="${%hostConnectionsDescr}">
+      		<div class="cpwr-section">
+	      		<f:repeatable var="hostConn" items="${instance.hostConnections}" add="${%addHostConnection}">      
+	      			<st:include page="HostConnection.jelly" class="com.compuware.jenkins.common.configuration.HostConnection"/>
+	      		</f:repeatable>
+      		</div>
+    	</f:entry>
+	</f:section>
 </j:jelly>
