+++ src/main/resources/com/compuware/jenkins/common/configuration/CpwrGlobalConfiguration/config.properties
@@ -1,11 +1,7 @@
 addHostConnection=Add Host Connection
-codePage=Code page
 deleteHostConnection=Delete Host Connection
 hostConnections=Host connections
 hostConnectionsDescr=List of host connections
-hostPort=Host:port
-loginCredentials=Login credentials
-name=Name
 title=Compuware Configurations
 topazWbCliDirectoryWindows=Windows Topaz Workbench CLI Home
 topazWbCliDirectoryLinux=Linux Topaz Workbench CLI Home
