+++ src/main/resources/com/compuware/jenkins/common/configuration/HostConnection/HostConnection.properties
@@ -1,0 +1,3 @@
+codePage=Code page
+hostPort=Host:port
+description=Description
