+++ src/main/resources/com/compuware/jenkins/common/configuration/HostConnection/help-codePage.html
