+++ src/main/resources/com/compuware/jenkins/common/configuration/Messages.properties
@@ -5,3 +5,5 @@ checkHostPortFormatError=The host:port must contain a single colon character.
 checkHostPortInvalidPortError=The port is not valid.
 checkHostPortMissingHostError=A host is required.
 checkHostPortMissingPortError=A port is required.
+
+displayNameHostConnection=Host connection
