+++ src/test/java/com/compuware/jenkins/common/configuration/CpwrGlobalConfigurationTest.java
@@ -20,7 +20,6 @@ import static org.junit.Assert.assertEquals;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.util.Map;
-import javax.servlet.ServletException;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -30,8 +29,6 @@ import hudson.Launcher;
 import hudson.Proc;
 import hudson.model.TaskListener;
 import hudson.remoting.Channel;
-import hudson.util.FormValidation;
-import hudson.util.FormValidation.Kind;
 
 /**
  * Class for testing the Compuware global configurations.
@@ -48,103 +45,6 @@ public class CpwrGlobalConfigurationTest
 	public void setUp()
 	{
 		m_globalConfig = new CpwrGlobalConfiguration();
-	}
-
-	/**
-	 * The the host port definition.
-	 */
-	@Test
-	public void hostPortTest()
-	{
-		// empty String
-		String input = "";
-		String expectedMsg = Messages.checkHostPortEmptyError();
-		validateHostPortErrorMessage("Expecting empty host port message.", input, expectedMsg);
-
-		// no host
-		input = ":1234";
-		expectedMsg = Messages.checkHostPortMissingHostError();
-		validateHostPortErrorMessage("Expecting no host message.", input, expectedMsg);
-
-		// no port
-		input = "abcd:";
-		expectedMsg = Messages.checkHostPortMissingPortError();
-		validateHostPortErrorMessage("Expecting no port message.", input, expectedMsg);
-
-		// invalid port
-		input = "abcd:aaa";
-		expectedMsg = Messages.checkHostPortInvalidPortError();
-		validateHostPortErrorMessage("Expecting invalid port message.", input, expectedMsg);
-
-		// no colon
-		input = "abcd1234";
-		expectedMsg = Messages.checkHostPortFormatError();
-		validateHostPortErrorMessage("Expecting missing colon message.", input, expectedMsg);
-
-		// multiple colons
-		input = "ab:cd1:234";
-		expectedMsg = Messages.checkHostPortFormatError();
-		validateHostPortErrorMessage("Expecting error for multiple colons.", input, expectedMsg);
-
-		// valid host:port
-		input = "abc:1234";
-		FormValidation validation = m_globalConfig.doCheckHostPort(input);
-		assertEquals("Expecting valid host port.", validation.kind, Kind.OK);
-	}
-
-	/**
-	 * Validate the host port error messages.
-	 * 
-	 * @param msg
-	 *            the msg to be displayed if the tests fails.
-	 * @param input
-	 *            the input value for the host port
-	 * @param expectedMsg
-	 *            the expected error message
-	 */
-	private void validateHostPortErrorMessage(String msg, String input, String expectedMsg)
-	{
-		FormValidation validation = m_globalConfig.doCheckHostPort(input);
-		String actualMsg = validation.getMessage();
-		System.out.println("Expected: " + expectedMsg);
-		System.out.println("Actual  : " + expectedMsg);
-		assertEquals(msg, expectedMsg, actualMsg);
-	}
-
-	/**
-	 * Test the host connection name.
-	 */
-	@Test
-	public void hostConnectionNameTest() throws IOException, ServletException
-	{
-		// empty String
-		String input = "";
-		String expectedMsg = Messages.checkConnectionNameEmptyError();
-		validateConnectionNameErrorMessage("Expecting empty connection name message.", input, expectedMsg);
-
-		input = "abc:1234";
-		FormValidation validation = m_globalConfig.doCheckConnectionName(input);
-		assertEquals("Expecting valid connection name.", validation.kind, Kind.OK);
-	}
-
-	/**
-	 * Validate the host connection name error messages.
-	 * 
-	 * @param msg
-	 *            the msg to be displayed if the tests fails.
-	 * @param input
-	 *            the input value for the host connection name
-	 * @param expectedMsg
-	 *            the expected error message
-	 */
-	private void validateConnectionNameErrorMessage(String msg, String input, String expectedMsg)
-			throws IOException, ServletException
-	{
-		FormValidation validation = m_globalConfig.doCheckConnectionName(input);
-		String actualMsg = validation.getMessage();
-		System.out.println("Expected: " + expectedMsg);
-		System.out.println("Actual  : " + expectedMsg);
-		assertEquals(msg, expectedMsg, actualMsg);
 	}
 
 	/**
