+++ src/test/java/com/compuware/jenkins/common/configuration/HostConnectionTest.java
@@ -1,0 +1,145 @@
+/**
+ * The MIT License (MIT)
+ * 
+ * Copyright (c) 2017 Compuware Corporation
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation
+ * files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy,
+ * modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following conditions: The above copyright notice and this permission notice
+ * shall be included in all copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE
+ * WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
+ * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package com.compuware.jenkins.common.configuration;
+
+import static org.junit.Assert.assertEquals;
+import java.io.IOException;
+import javax.servlet.ServletException;
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import org.jvnet.hudson.test.JenkinsRule;
+import com.compuware.jenkins.common.configuration.HostConnection.DescriptorImpl;
+import hudson.util.FormValidation;
+import hudson.util.FormValidation.Kind;
+
+/**
+ * Class for testing the Compuware global host connection configuration.
+ */
+@SuppressWarnings("nls")
+public class HostConnectionTest
+{
+	@Rule
+	public JenkinsRule j = new JenkinsRule();
+
+	private HostConnection m_globalHostConnectionConfig;
+
+	@Before
+	public void setUp()
+	{
+		m_globalHostConnectionConfig = new HostConnection("test", "cw01:1234", "1047", "1");
+	}
+
+	/**
+	 * The the host port definition.
+	 */
+	@Test
+	public void hostPortTest()
+	{
+		// empty String
+		String input = "";
+		String expectedMsg = Messages.checkHostPortEmptyError();
+		validateHostPortErrorMessage("Expecting empty host port message.", input, expectedMsg);
+
+		// no host
+		input = ":1234";
+		expectedMsg = Messages.checkHostPortMissingHostError();
+		validateHostPortErrorMessage("Expecting no host message.", input, expectedMsg);
+
+		// no port
+		input = "abcd:";
+		expectedMsg = Messages.checkHostPortMissingPortError();
+		validateHostPortErrorMessage("Expecting no port message.", input, expectedMsg);
+
+		// invalid port
+		input = "abcd:aaa";
+		expectedMsg = Messages.checkHostPortInvalidPortError();
+		validateHostPortErrorMessage("Expecting invalid port message.", input, expectedMsg);
+
+		// no colon
+		input = "abcd1234";
+		expectedMsg = Messages.checkHostPortFormatError();
+		validateHostPortErrorMessage("Expecting missing colon message.", input, expectedMsg);
+
+		// multiple colons
+		input = "ab:cd1:234";
+		expectedMsg = Messages.checkHostPortFormatError();
+		validateHostPortErrorMessage("Expecting error for multiple colons.", input, expectedMsg);
+
+		// valid host:port
+		input = "abc:1234";
+		FormValidation validation = ((DescriptorImpl) m_globalHostConnectionConfig.getDescriptor()).doCheckHostPort(input);
+		assertEquals("Expecting valid host port.", validation.kind, Kind.OK);
+	}
+
+	/**
+	 * Validate the host port error messages.
+	 * 
+	 * @param msg
+	 *            the msg to be displayed if the tests fails.
+	 * @param input
+	 *            the input value for the host port
+	 * @param expectedMsg
+	 *            the expected error message
+	 */
+	private void validateHostPortErrorMessage(String msg, String input, String expectedMsg)
+	{
+		FormValidation validation = ((DescriptorImpl) m_globalHostConnectionConfig.getDescriptor()).doCheckHostPort(input);
+		String actualMsg = validation.getMessage();
+		System.out.println("Expected: " + expectedMsg);
+		System.out.println("Actual  : " + expectedMsg);
+		assertEquals(msg, expectedMsg, actualMsg);
+	}
+
+	/**
+	 * Test the host connection name.
+	 */
+	@Test
+	public void hostConnectionNameTest() throws IOException, ServletException
+	{
+		// empty String
+		String input = "";
+		String expectedMsg = Messages.checkConnectionNameEmptyError();
+		validateConnectionNameErrorMessage("Expecting empty connection name message.", input, expectedMsg);
+
+		input = "abc:1234";
+		FormValidation validation = ((DescriptorImpl) m_globalHostConnectionConfig.getDescriptor())
+				.doCheckConnectionName(input);
+		assertEquals("Expecting valid connection name.", validation.kind, Kind.OK);
+	}
+
+	/**
+	 * Validate the host connection name error messages.
+	 * 
+	 * @param msg
+	 *            the msg to be displayed if the tests fails.
+	 * @param input
+	 *            the input value for the host connection name
+	 * @param expectedMsg
+	 *            the expected error message
+	 */
+	private void validateConnectionNameErrorMessage(String msg, String input, String expectedMsg)
+			throws IOException, ServletException
+	{
+		FormValidation validation = ((DescriptorImpl) m_globalHostConnectionConfig.getDescriptor())
+				.doCheckConnectionName(input);
+		String actualMsg = validation.getMessage();
+		System.out.println("Expected: " + expectedMsg);
+		System.out.println("Actual  : " + expectedMsg);
+		assertEquals(msg, expectedMsg, actualMsg);
+	}
+}
