+++ src/test/java/com/compuware/jenkins/common/utils/NumericStringComparatorTest.java
@@ -13,7 +13,6 @@ package com.compuware.jenkins.common.utils;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
-import org.junit.Before;
 import org.junit.Test;
 
 /**
